# openapi.yaml
openapi: 3.0.1
info:
  title: RonCode Plugin
  description: A plugin that allows the user to create, retrieve, and manage files using ChatGPT.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /files:
    post:
      operationId: add_file
      summary: Add a file to the list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                  description: The filename of the file to add to the list.
                file:
                  type: string
                  description: The content of the file to add to the list.
      responses:
        "200":
          description: OK
    get:
      operationId: get_files
      summary: Get the list of files
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        filename:
                          type: string
                          description: The filename of the file.
                        content:
                          type: string
                          description: The content of the file.
    delete:
      operationId: delete_file
      summary: Delete a file from the list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file_idx:
                  type: integer
                  description: The index of the file to delete.
      responses:
        "200":
          description: OK
  /upload:
    post:
      operationId: upload_files
      summary: Upload files from the UPLOAD folder to the scripts list and list the files
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        filename:
                          type: string
                          description: The filename of the file.
                        content:
                          type: string
                          description: The content of the file.
  /download:
    post:
      operationId: download_files
      summary: Download files from the scripts list to the DOWNLOAD folder
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                  description: The filename of the script to download. If not provided, all files will be downloaded.
      responses:
        "200":
          description: OK